+++ config/environments/development.rb
@@ -43,7 +43,7 @@ Rails.application.configure do
   config.action_mailer.perform_caching = false
 
   config.action_mailer.delivery_method = :smtp
-  config.action_mailer.smtp_settings = { address: '127.0.0.1', port: 1025 }
+  config.action_mailer.smtp_settings = { address: 'fabmanager-mailcatcher', port: 1025 }
   config.action_mailer.raise_delivery_errors = false
   config.action_mailer.default_url_options = {
     host: Rails.application.secrets.default_host,