+++ db/migrate/20140522180930_create_projects_components.rb
@@ -3,8 +3,8 @@
 class CreateProjectsComponents < ActiveRecord::Migration[4.2]
   def change
     create_table :projects_components do |t|
-    	t.belongs_to :project, index: true
-    	t.belongs_to :component, index: true
+      t.belongs_to :project, index: true
+      t.belongs_to :component, index: true
     end
   end
 end