+++ db/seeds.rb
@@ -278,7 +278,7 @@ end
 
 unless Setting.find_by(name: 'twitter_name').try(:value)
   setting = Setting.find_or_initialize_by(name: 'twitter_name')
-  setting.value = 'fab_manager'
+  setting.value = 'Fab_Manager'
   setting.save
 end