+++ setup/setup.sh
@@ -50,11 +50,6 @@ system_requirements()
       echo -e "\e[91m[ ❌ ] $_command was not found, exiting...\e[39m" && exit 1
     fi
   done
-  if ! command -v awk || ! [[ $(awk -W version) =~ ^GNU ]]
-  then
-    echo "Please install GNU Awk before running this script."
-    echo "\e[91m[ ❌ ] GNU awk was not found, exiting...\e[39m" && exit 1
-  fi
   printf "\e[92m[ ✔ ] All requirements successfully checked.\e[39m \n\n"
 }
 
@@ -111,6 +106,10 @@ prepare_files()
 {
   FABMANAGER_PATH=${1:-/apps/fabmanager}
 
+  echo -e "Fab-Manager will be installed in \e[31m$FABMANAGER_PATH\e[0m"
+  read -rp "Continue? (Y/n) " confirm </dev/tty
+  if [[ "$confirm" = "n" ]]; then exit 1; fi
+
   sudo mkdir -p "$FABMANAGER_PATH/config"
   sudo chown -R "$(whoami)" "$FABMANAGER_PATH"
 
@@ -147,6 +146,10 @@ prepare_files()
   \curl -sSL https://raw.githubusercontent.com/sleede/fab-manager/master/setup/docker-compose.yml > "$FABMANAGER_PATH/docker-compose.yml"
 }
 
+yq() {
+  docker run --rm -i -v "${FABMANAGER_PATH}:/workdir" mikefarah/yq yq "$@"
+}
+
 prepare_nginx()
 {
   if [ "$NGINX" != "n" ]; then
@@ -156,7 +159,18 @@ prepare_nginx()
     sed -i.bak "s/URL_WITH_PROTOCOL_HTTPS/https:\/\/${MAIN_DOMAIN[0]}/g" "$FABMANAGER_PATH/config/nginx/fabmanager.conf.ssl"
   else
     # if nginx is not installed, remove its associated block from docker-compose.yml
-    awk '$1 == "nginx:"{t=1; next};t==1 && /:[[:blank:]]*$/{t=0};t != 1' docker-compose.yml > "$FABMANAGER_PATH/.awktmpfile" && mv "$FABMANAGER_PATH/.awktmpfile" "$FABMANAGER_PATH/docker-compose.yml"
+    echo "Removing nginx..."
+    yq d -i docker-compose.yml services.nginx
+    read -rp "Do you want to map the Fab-manager's service to an external network? (Y/n) " confirm </dev/tty
+    if [ "$confirm" != "n" ]; then
+      echo "Adding a network configuration to the docker-compose.yml file..."
+      yq w -i docker-compose.yml networks.web.external true
+      yq w -i docker-compose.yml networks.db ''
+      yq w -i docker-compose.yml services.fabmanager.networks[+] web
+      yq w -i docker-compose.yml services.fabmanager.networks[+] db
+      yq w -i docker-compose.yml services.postgres.networks[+] db
+      yq w -i docker-compose.yml services.redis.networks[+] db
+    fi
   fi
 }
 
@@ -271,6 +285,7 @@ setup_assets_and_databases()
   printf "\n\nWe will now create the default administrator of Fab-manager.\n"
   read_email
   PASSWORD=$(read_password)
+  printf "\nOK. We will fulfill the database now...\n"
   cd "$FABMANAGER_PATH" && docker-compose run --rm -e ADMIN_EMAIL="$EMAIL" -e ADMIN_PASSWORD="$PASSWORD" fabmanager bundle exec rake db:seed # seed the database
 
   # now build the assets